% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{multiple variavble as numeric}
\usage{
as_numeric(data, col_name)
}
\arguments{
\item{data}{A data frame}

\item{col_name}{Mutliple character vectors (column name)}
}
\description{
Convert multiple variable (column) as numeric variable
}
\examples{
\dontrun{
DF <- data.frame("a" = as.character(0:5),
"col_a" = paste(0:5, ".1", sep = ""),
"col_b" = letters[1:6],
stringsAsFactors = FALSE)

test_01 <- as_numeric(data = DF, col_name = c("col_a", "col_b"))
}

}
