% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_stopwordsId.R
\name{clean_text}
\alias{clean_text}
\title{Delete Indonesian stopwords}
\usage{
clean_text(data_name, col_name)
}
\arguments{
\item{data_name}{A data frame name}

\item{col_name}{A column name of a data frame}
}
\value{
A normalised column in a data frame
}
\description{
Used to strip out stopwords bahasa Indoensia
}
\examples{
\dontrun{
library(cleanBahasa)
sample_text <- data.frame(text = c("di kantor saya sedang ada banyak kerjaan",
"kami akan pergi dengan mobil "))

test_01 <- clean_text(data_name = sample_text, col_name = "text")
}

}
