% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_text.R
\name{normal_text}
\alias{normal_text}
\title{normalise text}
\usage{
normal_text(data_name, col_name, nomral_lex)
}
\arguments{
\item{data_name}{A data frame name}

\item{col_name}{A column name of a data frame}

\item{nomral_lex}{A two column data frame contain of actual terms and its normal term}
}
\value{
A normalised column in a data frame
}
\description{
Used to nomalised text of a column in a data frame based on user defined
normalisation lexicon
}
\examples{
\dontrun{
library(cleanBahasa)
sample_text <- data.frame(text = c("ak ad kerjaan", "dia sdh pergi"))
kt_normal <- data.frame(from = c("ad", "ak", "sdh"),
                        to = c("ada", "aku", "sudah"))

test_01 <- normal_text(data_name = sample_text,
col_name = "text", nomral_lex = kt_normal)
}

}
