% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_group.R
\name{top_group}
\alias{top_group}
\title{Top n by group}
\usage{
top_group(data, groupx, columnx, n)
}
\arguments{
\item{data}{A data frame}

\item{groupx}{A Character group name of a column}

\item{columnx}{A Character column name to take}

\item{n}{Top n of a column to take}
}
\value{
Top n by group
}
\description{
Taking top n by group in a column
}
\examples{
\dontrun{
library(cleanBahasa)
data <- data.frame(
col_1   = runif(90),
col_2 = gl(3, 30))

data <- top_group(data, "col_1", "col_2", 5)
}

}
